\name{coleman.dat}
\docType{data}
\alias{coleman.dat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Coleman data }
\description{
  A good description of these data is coming in future releases
}
\usage{data(coleman.dat)}
\format{A data set}
\source{Coleman}
\references{Coleman}
\keyword{datasets}

