\name{roblm.Chi.prime.2}
\alias{roblm.Chi.prime.2}
\title{ Second derivative of Tukey's bi-square loss function }
\description{
}
\usage{
roblm.Chi.prime.2(x, cc)
}
\arguments{
  \item{x}{ argument }
  \item{cc}{ tuning constant }
}
\details{
}
\value{
}
\author{ Matias Salibian-Barrera }
\keyword{ robust }% at least one, from doc/KEYWORDS
