\name{roblm.Chi.prime}
\alias{roblm.Chi.prime}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Derivative of Tukey's bi-square loss function }
\description{
}
\usage{
roblm.Chi.prime(x, cc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ argument }
  \item{cc}{ tuning constant }
}
\details{
}
\value{
}
\author{ Matias Salibian-Barrera }
\keyword{ robust }% at least one, from doc/KEYWORDS
