\name{roblm.Chi}
\alias{roblm.Chi}
\title{ Tukey's bi-square loss function }
\description{
}
\usage{
roblm.Chi(x, cc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ argument }
  \item{cc}{ tuning constant }
}
\details{
}
\value{
}
\author{ Matias Salibian-Barrera }
\keyword{ robust }
