\name{roblm-package}
\alias{roblm-package}
\docType{package}
\title{
MM-regression estimators
}
\description{
MM-regression estimators
}
\details{
\tabular{ll}{
Package: \tab roblm\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2006-03-28\cr
License: \tab GPL 2.0 \cr
}

MM-regression estimators
}
\author{
Matias Salibian-Barrera

Maintainer: Matias Salibian-Barrera <matias@stat.ubc.ca>
}
\references{
Croux, C., Dhaene, G. and Hoorelbeke, D. (2003)
Robust standard errors
for robust estimators, Discussion Papers Series 03.16, K.U. Leuven, CES.

Rousseeuw, P.J. and Yohai, V.J. (1984)
Robust regression by means of 
S-estimators, In Robust and Nonlinear Time Series, J. Franke, W. 
H\"ardle and
R. D.
Martin (eds.). Lectures Notes in Statistics 26, 256-272, Springer Verlag, 
New York.

Salibian-Barrera, M. and Yohai, V.J. (2006) 
A fast algorithm for S-regression estimates, Journal of 
Computational and Graphical Statistics, in press.

Yohai, V.J. (1987)
High breakdown-point and high efficiency estimates for
regression, The Annals of Statistics 15, 642-65.
}
\keyword{ package }

