\name{roblm.Psi.prime}
\alias{roblm.Psi.prime}
\title{ Derivative of Tukey's bi-square score function}
\description{
}
\usage{
roblm.Psi.prime(x, cc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ argument }
  \item{cc}{ tuning constan }
}
\details{
}
\value{
}
\author{ Matias Salibian-Barrera }
\keyword{ robust }% at least one, from doc/KEYWORDS
