\name{roblm.Psi}
\alias{roblm.Psi}
\title{ Tukey's bi-square score function }
\description{
}
\usage{
roblm.Psi(x, cc)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ argument }
  \item{cc}{ tuning constant }
}
\details{
}
\value{
}
\author{ Matias Salibian-Barrera}
\keyword{ robust }% at least one, from doc/KEYWORDS
