\name{stackloss.dat}
\docType{data}
\alias{stackloss.dat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Stack Loss data }
\description{
  A good description of these data is coming in future releases
}
\usage{data(stackloss.dat)}
\format{A data set}
\source{Source}
\references{References}
\keyword{datasets}

